Messages delete option checkout - no warnings added)

+++ Rest/src/com/capgemini/rest/controller/MessagesController.java
@@ -166,7 +166,12 @@ public class MessagesController
 	
 	@RequestMapping(method = RequestMethod.POST, value = "/campaign-step-message-delete")
 	public String removeMessage(@RequestParam("messageId") long messageId, @RequestParam("stepId") long stepId, @RequestParam("campId") long campId, Model model) {
-		messageService.deleteMessageById(messageId);
+		
+		if(!messageService.getMessageById(messageId).isAlreadySent())
+		{
+			messageService.deleteMessageById(messageId);
+		}
+		
 		return "redirect:/service/campaign-step-messages/" + campId + "/" + stepId;
 	}
